feat: Implement storage migration summary and enhance project management UI
- Added a comprehensive storage migration summary document detailing recent changes, tests executed, potential risks, and mitigations.
- Introduced a new ProjectsTable component for displaying project details with actions.
- Created a ProjectOverviewPlaceholder component to provide a summary view for selected projects.
- Developed a CreateProjectWizard component with multi-step functionality for project creation, including project details, file selection, and review steps.
- Implemented validation logic for project creation forms and integrated file handling with support for multiple file types.
- Added utility functions for file descriptor management and validation of project details and files.
- Enhanced testing coverage for the CreateProjectWizard component to ensure proper functionality and error handling.
