+++ src/components/projects/ProjectsPanel.tsx
@@ -5,20 +5,12 @@ import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import { listProjects, type CreateProjectResponse, type ProjectListItem } from "@/ipc";
 
-import { CreateProjectWizard } from "./CreateProjectWizard";
+import { ProjectsTable, type ProjectManagerRow } from "./components/ProjectsTable";
+import { CreateProjectWizard } from "./wizard/CreateProjectWizard";
 
 type ProjectsPanelProps = {
   onOpenProject?: (project: ProjectListItem) => void;
 };
-
-interface ProjectManagerRow {
-  id: string;
-  name: string;
-  languagePair: string;
-  files: number;
-  updated: string;
-  status: string;
-}
 
 const DATE_FORMATTER = new Intl.DateTimeFormat(undefined, { dateStyle: "medium" });
 
@@ -40,9 +32,7 @@ export function ProjectsPanel({ onOpenProject }: ProjectsPanelProps = {}) {
       setProjects(response);
     } catch (unknownError) {
       const message =
-        unknownError instanceof Error
-          ? unknownError.message
-          : "Unable to load projects.";
+        unknownError instanceof Error ? unknownError.message : "Unable to load projects.";
       setError(message);
     } finally {
       if (options.showSpinner) {
@@ -67,18 +57,17 @@ export function ProjectsPanel({ onOpenProject }: ProjectsPanelProps = {}) {
     [onOpenProject, projects],
   );
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  function handleProjectCreated(_: CreateProjectResponse) {
+  const handleProjectCreated = useCallback((_: CreateProjectResponse) => {
     startRefresh(() => {
       void loadProjects();
     });
-  }
+  }, [loadProjects, startRefresh]);
 
-  function handleRefresh() {
+  const handleRefresh = useCallback(() => {
     startRefresh(() => {
       void loadProjects({ showSpinner: true });
     });
-  }
+  }, [loadProjects, startRefresh]);
 
   return (
     <section className="flex w-full flex-col gap-4 p-6" aria-labelledby="projects-heading">
@@ -132,79 +121,12 @@ export function ProjectsPanel({ onOpenProject }: ProjectsPanelProps = {}) {
             <p className="text-xs">Create your first project to get started.</p>
           </div>
         ) : (
-          <ProjectManagerTable rows={rows} onOpenProject={handleOpenProject} />
+          <ProjectsTable rows={rows} onOpenProject={handleOpenProject} />
         )}
       </div>
 
-      <CreateProjectWizard
-        open={isWizardOpen}
-        onOpenChange={setWizardOpen}
-        onProjectCreated={handleProjectCreated}
-      />
+      <CreateProjectWizard open={isWizardOpen} onOpenChange={setWizardOpen} onProjectCreated={handleProjectCreated} />
     </section>
-  );
-}
-
-function ProjectManagerTable({
-  rows,
-  onOpenProject,
-}: {
-  rows: ProjectManagerRow[];
-  onOpenProject?: (projectId: string) => void;
-}) {
-  return (
-    <div
-      className="space-y-1 overflow-hidden rounded-md border border-border/60 bg-background/80 shadow-sm"
-      role="table"
-      aria-label="Projects table"
-    >
-      <div
-        className="grid grid-cols-[minmax(0,1.5fr)_minmax(0,1.1fr)_72px_110px_minmax(0,1fr)_minmax(0,0.9fr)] gap-3 border-b border-border/60 bg-muted/40 px-3 py-2 text-[0.65rem] font-medium uppercase tracking-wide text-muted-foreground"
-        role="row"
-      >
-        <span role="columnheader">Project</span>
-        <span role="columnheader">Languages</span>
-        <span role="columnheader">Files</span>
-        <span role="columnheader">Updated</span>
-        <span role="columnheader">Status</span>
-        <span role="columnheader" className="text-right">
-          Actions
-        </span>
-      </div>
-      {rows.map((row) => (
-        <div
-          key={row.id}
-          className="grid grid-cols-[minmax(0,1.5fr)_minmax(0,1.1fr)_72px_110px_minmax(0,1fr)_minmax(0,0.9fr)] items-center gap-3 px-3 py-2 text-xs"
-          role="row"
-        >
-          <div className="truncate font-medium text-foreground" role="cell">
-            {row.name}
-          </div>
-          <div className="truncate text-muted-foreground" role="cell">
-            {row.languagePair}
-          </div>
-          <div className="text-muted-foreground" role="cell">
-            {row.files}
-          </div>
-          <div className="text-muted-foreground" role="cell">
-            {row.updated}
-          </div>
-          <div className="truncate text-foreground" role="cell">
-            {row.status}
-          </div>
-          <div className="flex justify-end" role="cell">
-            <Button
-              size="sm"
-              variant="outline"
-              onClick={() => onOpenProject?.(row.id)}
-              disabled={!onOpenProject}
-            >
-              Open Project
-            </Button>
-          </div>
-        </div>
-      ))}
-    </div>
   );
 }
